parser: simplify the generated tree by flattening repeats
If a repeated section or keyword is only mentioned once, don't make a
nested array of them which simplifies the structure and matches what is
currently being done in AiiDA for example.

The downside is that this requires knowledge of the CP2K schema when
generating a CP2K input from this structure again because we have to
know the datatype of a keyword to be able to figure out whether a given
array is due to repeated keywords or n_var != 1.